Centralized Pubx Initialization
